feat: Update setup script
- Fix a bug, copy wrong libhterrain
- Copy libraries to Godot project.

+++ scripts/setup.bash
@@ -2,6 +2,7 @@
 
 SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
 BUILD_DIR=${SCRIPT_DIR}/../build
+GAME_PROJECT_DIR=${SCRIPT_DIR}/../OwnWar
 
 mkdir -p ${BUILD_DIR}
 
@@ -11,14 +12,20 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-echo "Extracting Godot engine..."
-docker run --rm --entrypoint cat ownwar-deps-builder:latest \
-    /build/Godot_v3.3.5-modified_x11.64 >${BUILD_DIR}/Godot_v3.3.5-modified_x11.64
+if [ ! -f "${BUILD_DIR}/Godot_v3.3.5-modified_x11.64" ]; then
+    echo "Extracting Godot engine..."
+    docker run --rm --entrypoint cat ownwar-deps-builder:latest \
+        /build/Godot_v3.3.5-modified_x11.64 >${BUILD_DIR}/Godot_v3.3.5-modified_x11.64
+else
+    echo "Godot engine exists."
+fi
 
 echo "Extracting native libraries..."
 docker run --rm --entrypoint cat ownwar-deps-builder:latest \
     /build/libgd_3d_batcher.so >${BUILD_DIR}/libgd_3d_batcher.so
 docker run --rm --entrypoint cat ownwar-deps-builder:latest \
-    /build/libgd_3d_batcher.so >${BUILD_DIR}/libhterrain.so
+    /build/libhterrain.so >${BUILD_DIR}/libhterrain.so
 docker run --rm --entrypoint cat ownwar-deps-builder:latest \
     /build/libownwar.so >${BUILD_DIR}/libownwar.so
+
+cp ${BUILD_DIR}/*.so ${GAME_PROJECT_DIR}/gd/lib/.
